Consolidated valuechecks. Moved isNaN to drakeFloatingPointUtil and changed the return type to boolean.
