+++ util/drakeUtil.h
@@ -6,17 +6,10 @@
  */
 
 #include "mex.h"
-#include <string>
 #include <stdexcept>
 #include <vector>
 #include <utility>
 #include <Eigen/Core>
-
-// needed only for valuecheck (but keeping that code in the header avoids explicit instantiations)
-#include <sstream> 
-#include <cmath>
-// end valuecheck includes
-
 
 #ifndef DRAKE_UTIL_H_
 #define DRAKE_UTIL_H_
@@ -115,43 +108,5 @@ DLLEXPORT mxArray* mxGetFieldSafe(const mxArray* array, size_t index, std::strin
 DLLEXPORT void mxSetFieldSafe(mxArray* array, size_t index, std::string const & fieldname, mxArray* data);
 
 DLLEXPORT void sizecheck(const mxArray* mat, int M, int N);
-
-template<typename Derived>
-std::string to_string(const Eigen::MatrixBase<Derived> & a)
-{
-	std::stringstream ss;
-	ss << a;
-	return ss.str();
-}
-
-inline int my_isnan(double x) {
-#ifdef WIN32
-  return _isnan(x);
-#else
-  return std::isnan(x);
-#endif
-}
-
-template<typename DerivedA, typename DerivedB>
-void valuecheck(const Eigen::MatrixBase<DerivedA>& a, const Eigen::MatrixBase<DerivedB>& b, double tol, std::string error_msg)
-{
-  // note: isApprox uses the L2 norm, so is bad for comparing against zero
-  if (a.rows() != b.rows() || a.cols() != b.cols()) {
-    throw std::runtime_error("Drake:ValueCheck ERROR:" + error_msg + "size mismatch: (" + std::to_string(static_cast<unsigned long long>(a.rows())) + " by " + std::to_string(static_cast<unsigned long long>(a.cols())) + ") and (" + std::to_string(static_cast<unsigned long long>(b.rows())) + " by " + std::to_string(static_cast<unsigned long long>(b.cols())) + ")");
-  }
-  if (!(a-b).isZero(tol)) {
-    if (!a.allFinite() && !b.allFinite()) {
-      // could be failing because inf-inf = nan
-      bool ok=true;
-      for (int i=0; i<a.rows(); i++)
-        for (int j=0; j<a.cols(); j++) {
-          ok = ok && ((a(i,j) == std::numeric_limits<double>::infinity() && b(i,j) == std::numeric_limits<double>::infinity()) || (a(i,j) == -std::numeric_limits<double>::infinity() && b(i,j) == -std::numeric_limits<double>::infinity()) || (my_isnan(a(i,j)) && my_isnan(b(i,j))) || (std::abs(a(i,j)-b(i,j))<tol));
-        }
-      if (ok) return;
-    }
-    error_msg += "A:\n" + to_string(a) + "\nB:\n" + to_string(b) + "\n";
-    throw std::runtime_error("Drake:ValueCheck ERROR:" + error_msg);
-  }
-}
 
 #endif /* DRAKE_UTIL_H_ */
