+++ util/test/testDrakeGeometryUtil.cpp
@@ -200,7 +200,7 @@ void testSpatialCrossProduct()
   auto b = (Matrix<double, TWIST_SIZE, TWIST_SIZE>::Identity()).eval();
   auto a_crm_b = crossSpatialMotion(a, b);
   auto a_crf_b = crossSpatialForce(a, b);
-  valuecheck(a_crf_b, -a_crm_b.transpose());
+  valuecheck(a_crf_b, -a_crm_b.transpose(), 1e-8);
 }
 
 void testdrpy2rotmat()
