+++ util/test/testUtil.h
@@ -6,9 +6,11 @@
 #endif
 
 #include <Eigen/Core>
-#include <iostream>
+#include <sstream>
+#include <string>
 #include <stdexcept>
 #include <cmath>
+#include "../drakeFloatingPointUtil.h"
 
 // requires <chrono>, which isn't available in MSVC2010...
 #if !defined(WIN32) && !defined(WIN64)
@@ -31,17 +33,43 @@ struct measure
 };
 #endif
 
-template<typename DerivedA, typename DerivedB>
-void valuecheck(const Eigen::DenseBase<DerivedA>& a, const Eigen::DenseBase<DerivedB>& b, typename DerivedA::Scalar tolerance = 1e-8)
+template<typename Derived>
+std::string to_string(const Eigen::MatrixBase<Derived> & a)
 {
-  if (!a.isApprox(b)) {
-    std::ostringstream stream;
-    stream << "Expected:\n" << a << "\nbut got:\n" << b << "\n";
-    throw std::runtime_error(stream.str());
+  std::stringstream ss;
+  ss << a;
+  return ss.str();
+}
+
+template<typename DerivedA, typename DerivedB>
+void valuecheck(const Eigen::MatrixBase<DerivedA>& a, const Eigen::MatrixBase<DerivedB>& b, double tol, std::string error_msg = "")
+{
+  // note: isApprox uses the L2 norm, so is bad for comparing against zero
+  if (a.rows() != b.rows() || a.cols() != b.cols()) {
+    throw std::runtime_error(
+        "Drake:ValueCheck ERROR:" + error_msg + "size mismatch: (" + std::to_string(static_cast<unsigned long long>(a.rows())) + " by " + std::to_string(static_cast<unsigned long long>(a.cols())) + ") and (" + std::to_string(static_cast<unsigned long long>(b.rows())) + " by "
+            + std::to_string(static_cast<unsigned long long>(b.cols())) + ")");
+  }
+  if (!(a - b).isZero(tol)) {
+    if (!a.allFinite() && !b.allFinite()) {
+      // could be failing because inf-inf = nan
+      bool ok = true;
+      for (int i = 0; i < a.rows(); i++)
+        for (int j = 0; j < a.cols(); j++) {
+          bool both_positive_infinity = a(i, j) == std::numeric_limits<double>::infinity() && b(i, j) == std::numeric_limits<double>::infinity();
+          bool both_negative_infinity = a(i, j) == -std::numeric_limits<double>::infinity() && b(i, j) == -std::numeric_limits<double>::infinity();
+          bool both_nan = isNaN(a(i, j)) && isNaN(b(i, j));
+          ok = ok && (both_positive_infinity || both_negative_infinity || (both_nan) || (std::abs(a(i, j) - b(i, j)) < tol));
+        }
+      if (ok)
+        return;
+    }
+    error_msg += "A:\n" + to_string(a) + "\nB:\n" + to_string(b) + "\n";
+    throw std::runtime_error("Drake:ValueCheck ERROR:" + error_msg);
   }
 }
 
-void valuecheck(double a, double b, double tolerance = 1e-8)
+void valuecheck(double a, double b, double tolerance)
 {
   if (std::abs(a - b) > tolerance) {
     std::ostringstream stream;
